packages/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "servingcard"
version = "0.1.0"
description = "Hardware-specific LLM serving configurations — model cards for serving"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "zenprocess" },
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "typer>=0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[project.scripts]
servingcard = "servingcard.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["servingcard"]
